+++ components/pages/submission-system/program-dashboard/ClinicalChart/RangeControlBar.tsx
