+++ components/pages/submission-system/program-dashboard/ClinicalChart/mockData.ts
